aeternus/lumina/requirements.txt
torch>=2.1.0
transformers>=4.35.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
pyyaml>=6.0
einops>=0.7.0
rotary-embedding-torch>=0.3.0
# flash-attn>=2.3.0  # optional, requires CUDA toolkit
datasets>=2.14.0
tokenizers>=0.15.0
tqdm>=4.65.0
wandb>=0.16.0
scikit-learn>=1.3.0
